+++ bank-feeds/src/codatbankfeeds/models/operations/create_source_account.py
@@ -4,27 +4,29 @@ from __future__ import annotations
 import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
+from ..shared import sourceaccount as shared_sourceaccount
 from typing import Optional
 
 
 
 @dataclasses.dataclass
-class DeleteBankFeedBankAccountRequest:
-    account_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'accountId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an account"""
+class CreateSourceAccountRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    source_account: Optional[shared_sourceaccount.SourceAccount] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     
 
 
 
 
 @dataclasses.dataclass
-class DeleteBankFeedBankAccountResponse:
+class CreateSourceAccountResponse:
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    source_account: Optional[shared_sourceaccount.SourceAccount] = dataclasses.field(default=None)
+    r"""Success"""