+++ bank-feeds/RELEASES.md
@@ -709,3 +709,13 @@ Based on:
 - [python v0.35.1] bank-feeds
 ### Releases
 - [PyPI v0.35.1] https://pypi.org/project/codat-bankfeeds/0.35.1 - bank-feeds
+
+## 2023-09-05 00:13:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.35.2] bank-feeds
+### Releases
+- [PyPI v0.35.2] https://pypi.org/project/codat-bankfeeds/0.35.2 - bank-feeds